% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula-class.R
\docType{class}
\name{copulagev}
\alias{copulagev}
\title{Classe Copulas GEV}
\description{
The class provides tools to define the type and dependency parameter attributes of a copula, and the associated methods such as its value, conditional probability and failure probability
}
\details{
Classe Copulas GEV
}
\section{Methods}{

\describe{
 \item{\code{new(cop.type, Spearman.rho)}}{This method is used to create object of this class with \code{cop.type} as copula class and \code{Spearman.rho} the Spearmon's dependency index value.}
 \item{\code{initialize(cop.type, Spearman.rho)}}{This method is used to create object of this class with \code{cop.type} as copula class and \code{Spearman.rho} the Spearmon's dependency index value.}
 \item{\code{set.class.cop(copula.name)}}{This method is used to set da copula features using \code{copula.name}.}
 \item{\code{get.class.cop()}}{This method is used to create object of this class with  Spearmon's dependency index value.}}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{copula.type}}{An extreme-value copula class within gumbel, galambos, huslerReiss, tawn,}

\item{\code{rho}}{The value of Spearman's dependence index}

\item{\code{theta.u1}}{Parameter Vector}

\item{\code{theta.u2}}{Parameter Vector}

\item{\code{name}}{Name for test}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-copulagev-new}{\code{copulagev$new()}}
\item \href{#method-copulagev-set.class.cop}{\code{copulagev$set.class.cop()}}
\item \href{#method-copulagev-get.class.cop}{\code{copulagev$get.class.cop()}}
\item \href{#method-copulagev-assembly}{\code{copulagev$assembly()}}
\item \href{#method-copulagev-clone}{\code{copulagev$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copulagev-new"></a>}}
\if{latex}{\out{\hypertarget{method-copulagev-new}{}}}
\subsection{Method \code{new()}}{
Create an object of copulagev class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copulagev$new(cop.type = NULL, Spearman.rho = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cop.type}}{Name of the copula}

\item{\code{Spearman.rho}}{Coeficient of Spearman correlation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copulagev-set.class.cop"></a>}}
\if{latex}{\out{\hypertarget{method-copulagev-set.class.cop}{}}}
\subsection{Method \code{set.class.cop()}}{
Set the main features of a copula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copulagev$set.class.cop(copula.name, rho.coef)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cop.type}}{Name of the copula}

\item{\code{Spearman.rho}}{Coeficient of Spearman correlation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copulagev-get.class.cop"></a>}}
\if{latex}{\out{\hypertarget{method-copulagev-get.class.cop}{}}}
\subsection{Method \code{get.class.cop()}}{
Get the main features of a copula
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copulagev$get.class.cop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copulagev-assembly"></a>}}
\if{latex}{\out{\hypertarget{method-copulagev-assembly}{}}}
\subsection{Method \code{assembly()}}{
Assembly all copula data including the type of correlation, the value of dependence coefficient, marginals parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copulagev$assembly()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copulagev-clone"></a>}}
\if{latex}{\out{\hypertarget{method-copulagev-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{copulagev$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
