% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{failure.tbev.2}
\alias{failure.tbev.2}
\title{The function computes the probability of failure given by P(X<Y), where X~F and Y~F are random variables with a GEV distribution with parameter location mu, scale sigma and shape xi. It has a special importance in reliability engineering where X and Y represent stress and strength, respectively.}
\usage{
failure.tbev.2(r, theta1, theta2)
}
\arguments{
\item{r}{represents the parameter of dependence function.}

\item{theta1}{Vector of parameter (xi,mu,sigma) of the distribuicao 1 F_X}

\item{theta2}{Vector of parameter (xi,mu,sigma) of the distribuicao 2 F_Y}
}
\value{
The value of the probability failure
}
\description{
The function computes the probability of failure given by P(X<Y), where X~F and Y~F are random variables with a GEV distribution with parameter location mu, scale sigma and shape xi. It has a special importance in reliability engineering where X and Y represent stress and strength, respectively.
}
