% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{tbev.interval}
\alias{tbev.interval}
\title{Verifica o intervalo do suporte da distribuição GEV para cada componente}
\usage{
tbev.interval(theta1, theta2)
}
\arguments{
\item{theta1}{Vetor de parametros (xi,mu,sigma2) da distribuicao 1}

\item{theta2}{Vetor de parametros (xi,mu,sigma2) da distribuicao 2}
}
\value{
gives the support of distribution according to parameters
}
\description{
Verifica o intervalo do suporte da distribuição GEV para cada componente
}
