% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{transf.BEV}
\alias{transf.BEV}
\title{The transformation function that gives the name transformed BEV distribution}
\usage{
transf.BEV(x, mu, delta)
}
\arguments{
\item{x}{input of random variable to transformed.}

\item{mu}{parameter location}

\item{delta}{distortion parameter}
}
\value{
The value of the probability failure
}
\description{
The transformation function that gives the name transformed BEV distribution
}
